Added an antivirus enumeration for Windows
