+++ pwncat/modules/windows/manage/powershell/import.py
@@ -1,11 +1,10 @@
 #!/usr/bin/env python3
 import os
-from io import IOBase, BytesIO
+from io import BytesIO, IOBase
 from pathlib import Path
 
 import requests
-
-from pwncat.modules import Bool, Argument, BaseModule, ModuleFailed
+from pwncat.modules import Argument, BaseModule, Bool, ModuleFailed
 from pwncat.platform.windows import Windows
 
 
@@ -33,14 +32,14 @@ class Module(BaseModule):
         self.imported_modules = []
 
     def resolve_psmodule(self, session: "pwncat.manager.Session", path: str):
-        """ Resolve a module name into a file-like object """
+        """Resolve a module name into a file-like object"""
 
         if path.startswith("http://") or path.startswith("https://"):
             # Load from a URL
             r = requests.get(path, stream=True)
             if r.status_code != 200:
                 raise PSModuleNotFoundError(path)
-            return path.split("/")[-1], BytesIO(r.content)
+            return path.split("/")[-1], BytesIO(r.content + b"\n")
 
         orig_path = path
         path = Path(path)
@@ -62,7 +61,7 @@ class Module(BaseModule):
             if r.status_code != 200:
                 raise PSModuleNotFoundError(orig_path)
 
-            return (path.name, BytesIO(r.content + "\n"))
+            return (path.name, BytesIO(r.content + b"\n"))
         else:
             raise PSModuleNotFoundError(orig_path)
